+++ CMakeLists.txt
@@ -172,10 +172,10 @@ set(STLINK_HEADERS
         src/stlink-lib/chipid.h
         src/stlink-lib/commands.h
         src/stlink-lib/common_flash.h
-        src/stlink-lib/common.h
         src/stlink-lib/flash_loader.h
         src/stlink-lib/helper.h
         src/stlink-lib/libusb_settings.h
+        src/stlink-lib/lib_md5.h
         src/stlink-lib/logging.h
         src/stlink-lib/map_file.h
         src/stlink-lib/md5.h
@@ -194,6 +194,7 @@ set(STLINK_SOURCE
         src/stlink-lib/helper.c
         src/stlink-lib/logging.c
         src/stlink-lib/map_file.c
+        src/stlink-lib/lib_md5.c
         src/stlink-lib/md5.c
         src/stlink-lib/option_bytes.c
         src/stlink-lib/read_write.c
