+++ src/stlink-lib/helper.c
@@ -1,20 +1,21 @@
+/* == nightwalker-87: TODO: CONTENT AND USE OF THIS SOURCE FILE IS TO BE VERIFIED (07.06.2023) == */
+
 /*
  * File: helper.c
  *
  * General helper functions
  */
-
-#include <stddef.h>
-#include <stdint.h>
-#include <stdlib.h>
-
-#include "helper.h"
-
 #ifdef STLINK_HAVE_SYS_TIME_H
 #include <sys/time.h>
 #else
 #include <sys_time.h>
 #endif
+
+#include <stdint.h>
+#include <stddef.h>
+#include <stdlib.h>
+
+#include "helper.h"
 
 uint32_t time_ms() {
     struct timeval tv;
