+++ src/stlink-lib/map_file.c
@@ -4,23 +4,61 @@
  * File mapping
  */
 
-#include <fcntl.h>
 #include <stdint.h>
+#include <stdio.h>
+#include <string.h>
+
+#include <fcntl.h>
 #include <unistd.h>
 #include <sys/stat.h>
 
 #include <stlink.h>
-#include "logging.h"
 #include "map_file.h"
-#include "md5.h"
 
 #ifndef O_BINARY
 #define O_BINARY 0
 #endif
 
+// 1 GB max file size
 #ifndef MAX_FILE_SIZE
-#define MAX_FILE_SIZE (1<<20) // 1 GB max file size
+#define MAX_FILE_SIZE (1<<20)
 #endif
+
+/* Limit the block size to compare to 0x1800 as anything larger will stall the
+ * STLINK2 Maybe STLINK V1 needs smaller value!
+ */
+int32_t check_file(stlink_t *sl, mapped_file_t *mf, stm32_addr_t addr) {
+  size_t off;
+  size_t n_cmp = sl->flash_pgsz;
+
+  if (n_cmp > 0x1800) {
+    n_cmp = 0x1800;
+  }
+
+  for (off = 0; off < mf->len; off += n_cmp) {
+    size_t aligned_size;
+
+    size_t cmp_size = n_cmp; // adjust last page size
+
+    if ((off + n_cmp) > mf->len) {
+      cmp_size = mf->len - off;
+    }
+
+    aligned_size = cmp_size;
+
+    if (aligned_size & (4 - 1)) {
+      aligned_size = (cmp_size + 4) & ~(4 - 1);
+    }
+
+    stlink_read_mem32(sl, addr + (uint32_t)off, (uint16_t)aligned_size);
+
+    if (memcmp(sl->q_buf, mf->base + off, cmp_size)) {
+      return (-1);
+    }
+  }
+
+  return (0);
+}
 
 int32_t map_file(mapped_file_t *mf, const char *path) {
   int32_t error = -1;
